+++ RELEASES.md
@@ -461,3 +461,13 @@ Based on:
 - [typescript v0.1.0-alpha.42] .
 ### Releases
 - [NPM v0.1.0-alpha.42] https://www.npmjs.com/package/syllable-sdk/v/0.1.0-alpha.42 - .
+
+## 2025-02-17 20:05:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.493.2 (2.512.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.1.0-alpha.43] .
+### Releases
+- [NPM v0.1.0-alpha.43] https://www.npmjs.com/package/syllable-sdk/v/0.1.0-alpha.43 - .